OCPBUGS-30537: CVE-2024-24786 Bump google.golang.org/protobuf to v1.33.0
This PR bumps the version of google.golang.org/protobuf to v1.33.0 fix a
potential vulnerability in the protojson.Unmarhsl function [1] that can
occure when unmarshaling a message with a protobuf value.

Even if the function isn't used directly in Thanos it would be safer to
just bump it directly.

[1] https://pkg.go.dev/vuln/GO-2024-2611

(cherry picked from upstream PR https://github.com/thanos-io/thanos/pull/7191)
Signed-off-by: Daniel Mellado <dmellado@redhat.com>